product-service/requirements.txt
fastapi
uvicorn
requests
httpx
python-jose[cryptography]